Import of MUMIAW/File-FilterFuncs-0.51 from CPAN.
gitpan-cpan-distribution: File-FilterFuncs
gitpan-cpan-version:      0.51
gitpan-cpan-path:         MUMIAW/File-FilterFuncs-0.51.tar.gz
gitpan-cpan-author:       MUMIAW
gitpan-cpan-maturity:     released

+++ t/FF_Common.pm
@@ -2,35 +2,68 @@ package t::FF_Common;
 use strict;
 use warnings;
 use POSIX qw(tmpnam);
-use Exporter qw(import);
+use Exporter ();
 use File::Spec::Functions;
 
 BEGIN {
 	our @ISA = qw(Exporter);
-	our @EXPORT = qw(%Common);
+	our @EXPORT = qw(%Common slurp_file unslurp_file);
+	our @EXPORT_OK = qw(%Common);
 }
 
-my $tmpnam = tmpnam();
-# my $tmpnam = '/tmp/ff.test.dir';
 
-our %Common = (
-	tempdir => $tmpnam,
-	tempin => catfile($tmpnam,'input'),
-	tempout => catfile($tmpnam,'output'),
-);
+our $DEBUG;
+our %Common;
+
 
 sub init {
+	if ("@_" =~ /\bdebug\b/) {
+		$DEBUG = 1;
+	}
+
+	my $tmpnam = $DEBUG ? '/tmp/ff.test.dir' : tmpnam();
+	%Common = (
+		tempdir => $tmpnam,
+		tempin => catfile($tmpnam,'input'),
+		tempout => catfile($tmpnam,'output'),
+	);
+
 	return if (-d $Common{tempdir});
 	mkdir $Common{tempdir};
 }
 
 
+
 sub cleanup {
+	return if $DEBUG;
+
 	unlink $Common{tempin};
 	unlink $Common{tempout};
 	my @temps = glob(catfile($Common{tempdir},'t.*'));
 	unlink @temps;
 	rmdir $Common{tempdir};
+}
+
+sub slurp_file {
+	my $filename = shift;
+	my $fh;
+	local $/;
+
+	open $fh, '<:raw', $filename;
+	my $data = <$fh>;
+	close $fh;
+	$data;
+}
+
+sub unslurp_file {
+	my $filename = shift;
+	my $data = join('',@_);
+	my $fh;
+
+	open $fh, '>:raw', $filename;
+	print $fh $data;
+	close $fh;
+	1;
 }
